ASoC: wcd9xxx-mbhc: Reduce detection time for PLUG_TYPE_HEADSET
In lucky cases where a good headset is plugged in, detection latency can be
significantly reduced by setting the headset plug type early on, before entering
the correct-plug worker.

If there is a faulty detection and an inserted plug is detected as a headset early-on,
even though it is not, we will just fall back to the correct-plug worker to fix the detection error.

Signed-off-by: Sultanxda <sultanxda@gmail.com>